refactor: xmodule/library_content_module.py -> xmodule/library_content_block.py
