+++ lms/djangoapps/course_blocks/transformers/library_content.py
@@ -14,7 +14,7 @@ from openedx.core.djangoapps.content.block_structure.transformer import (
     BlockStructureTransformer,
     FilteringTransformerMixin
 )
-from xmodule.library_content_module import LibraryContentBlock  # lint-amnesty, pylint: disable=wrong-import-order
+from xmodule.library_content_block import LibraryContentBlock  # lint-amnesty, pylint: disable=wrong-import-order
 from xmodule.modulestore.django import modulestore  # lint-amnesty, pylint: disable=wrong-import-order
 
 from ..utils import get_student_module_as_dict
