+++ setup.py
@@ -21,7 +21,7 @@ XBLOCKS = [
     "html = xmodule.html_block:HtmlBlock",
     "image = xmodule.template_module:TranslateCustomTagBlock",
     "library = xmodule.library_root_xblock:LibraryRoot",
-    "library_content = xmodule.library_content_module:LibraryContentBlock",
+    "library_content = xmodule.library_content_block:LibraryContentBlock",
     "library_sourced = xmodule.library_sourced_block:LibrarySourcedBlock",
     "lti = xmodule.lti_module:LTIBlock",
     "poll_question = xmodule.poll_module:PollBlock",
